DateField + PopupJS/Html + Storage
Added a dateField on OdooPage to chose the next activity date and updated script to manage this information.
Revamped the Pop HTML to match new requirement as JS behind to generate new activity field and remove button.
Added the storage functionnality where PopupJS send Message to BackGround, that stores the object in local storage.
Added to PopupJS the functionnality that get the object from local storage.
Need to display the process on the Popup HTML + share object with script to customize script operations